Feature/ep 0005 (#6)
* EP-0005 - Renamed Rest Controllers and new Thymeleaf Controllers

* EP-0005 - Added Mapping annotations and Fetch type

* EP-0005 - Changed ticket info

* EP-0005 - Added base UI and Error handling

* EP-0005 - Modified Controllers, Added update and create pages and added delete buttons

* EP-0005 - Modified Ticket UI using new TicketView

* EP-0005 - Added deleting business logic

* EP-0005 - Added plane registrationNumber field

* EP-0005 - Added planeDto registrationNumber field

* EP-0005 - Added flight plane select and save

* EP-0005 - Added booking user select and save

* EP-0005 - Added plane creating page and view page registrationNumber

* EP-0005 - Added create user password

* EP-0005 - Added ticket view plane model

* EP-0005 - Added bookings ticket create, ticket creation change

+++ src/main/java/edu/grsu/practice/practice/service/impl/FlightServiceImpl.java
@@ -2,9 +2,12 @@ package edu.grsu.practice.practice.service.impl;
 
 import edu.grsu.practice.practice.dto.FlightDto;
 import edu.grsu.practice.practice.mapper.FlightMapper;
+import edu.grsu.practice.practice.mapper.PlaneMapper;
 import edu.grsu.practice.practice.model.Flight;
+import edu.grsu.practice.practice.model.Plane;
 import edu.grsu.practice.practice.repository.FlightRepository;
 import edu.grsu.practice.practice.service.FlightService;
+import edu.grsu.practice.practice.service.PlaneService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -20,11 +23,14 @@ public class FlightServiceImpl implements FlightService {
 
     public FlightMapper flightMapper;
     public FlightRepository flightRepository;
+    public PlaneService planeService;
+    public PlaneMapper planeMapper;
 
     @Override
     public FlightDto addFlight(FlightDto flightDto) {
         log.info("Adding flight: {}", flightDto);
         Flight flight = flightMapper.toEntity(flightDto);
+        flight.setPlane(planeMapper.toEntity(planeService.getPlane(flightDto.getPlaneId())));
         flightRepository.save(flight);
         return flightMapper.toDto(flight);
     }
@@ -41,6 +47,8 @@ public class FlightServiceImpl implements FlightService {
         log.info("Getting flight: {}", flightId);
         Optional<Flight> flightOptional = flightRepository.findById(flightId);
         Flight flight = flightOptional.orElseThrow();
+        Plane plane = planeMapper.toEntity(planeService.getPlane(flight.getPlane().getId()));
+        flight.setPlane(plane);
         return flightMapper.toDto(flight);
     }
 
@@ -49,6 +57,9 @@ public class FlightServiceImpl implements FlightService {
         log.info("Deleting flight: {}", flightId);
         Optional<Flight> flightOptional = flightRepository.findById(flightId);
         Flight flight = flightOptional.orElseThrow();
+        if (!flight.getTickets().isEmpty()) {
+            return false;
+        }
         flightRepository.delete(flight);
         return true;
     }
